+++ src/main/java/io/knotx/commons/cache/CacheFactory.java
@@ -13,23 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.knotx.commons.cache;
 
-rootProject.name = "knotx-commons"
+import io.vertx.core.Vertx;
+import io.vertx.core.json.JsonObject;
 
-pluginManagement {
-    val version: String by settings
-    plugins {
-        id("io.knotx.java-library") version version
-        id("io.knotx.codegen") version version
-        id("io.knotx.unit-test") version version
-        id("io.knotx.jacoco") version version
-        id("io.knotx.maven-publish") version version
-        id("io.knotx.release-java") version version
-        id("org.nosphere.apache.rat") version "0.6.0"
-    }
-    repositories {
-        mavenLocal()
-        jcenter()
-        gradlePluginPortal()
-    }
+public interface CacheFactory {
+
+  /**
+   * Type of this factory that will be referenced by clients
+   * @return type of this factory
+   */
+  String getType();
+
+  /**
+   * Create a Cache implementation using provided configuration.
+   *
+   * @param config configuration
+   * @param vertx Vert.x instance
+   * @return cache implementation
+   */
+  Cache create(JsonObject config, Vertx vertx);
+
 }